Extract src/actions.ts — the app-level verbs
The last block of the bootstrap trim: the small verbs several surfaces trigger
and none of them owns. A palette row, a context-menu row, a settings control
and a header button can all pin a project, change the sort, flip the theme or
answer a permission, so these sit below all of them.

Moved (four guarded ranges): addProject / addProjectPath / removeFavorite,
openProjectFolder, resolvePermission, setSort / cycleSort, toggleRail /
toggleInsp, effectiveTheme / setTheme / toggleTheme, and the app-level
setWtGroup with its episkoWtGroup console affordance.

They are one shape, and that shape is the reason they are not in state.ts:
**mutate the persisted preference, write it, repaint.** state.ts's setters
assign and nothing else — that is PLAN's own setX-per-variable decision — so
the persistence and the repaint had to live a layer up. Now they live in one
layer instead of scattered through main.ts. One hook, setActionsRenderAll.

Verified mechanical: line-sorted against `git show HEAD:src/main.ts` the
relocated block is identical modulo the export keywords. Cycle check: no
cycles across 34 modules.

main.ts 693 -> 636. Gates: tsc --noEmit clean (exit 0), 368 vitest pass. No
Rust touched.

Found while doing this, and fixed next in its own commit: the settings
window's worktree-grouping control calls state.ts's raw setWtGroup, which
assigns and nothing else — so the choice is never persisted and the sidebar
never regroups. Flagged here, not fixed here.
